Refactored tests to have one test for each case
