+++ src/Core/Types.Filters.Tests/QueryableFilterTests.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using HotChocolate.Execution;
 using Snapshooter.Xunit;
 using Xunit;
@@ -240,6 +242,65 @@ namespace HotChocolate.Types.Filters
 
             // assert
             result.MatchSnapshot();
+        }
+
+        [Fact]
+        public async Task Execute_DateTime_Filter()
+        {
+            // arrange
+            ISchema schema = SchemaBuilder.New()
+                .AddQueryType<QueryFooDateTime>(d => d
+                    .Name("Query")
+                    .Field(y => y.Foo)
+                    .UseFiltering())
+                .Create();
+
+            IQueryExecutor executor = schema.MakeExecutable();
+
+            IReadOnlyQueryRequest request = QueryRequestBuilder.New()
+                .SetQuery("{ foo(where: { foo_gte: \"2019-06-01\"}) {foo}}")
+                .Create();
+
+            // act
+            IExecutionResult result = await executor.ExecuteAsync(request);
+
+            // assert
+            result.MatchSnapshot();
+        }
+
+        [Fact]
+        public async Task Execute_DateTime_Filter_With_Variables()
+        {
+            // arrange
+            ISchema schema = SchemaBuilder.New()
+                .AddQueryType<QueryFooDateTime>(d => d
+                    .Name("Query")
+                    .Field(y => y.Foo)
+                    .UseFiltering())
+                .Create();
+
+            IQueryExecutor executor = schema.MakeExecutable();
+
+            IReadOnlyQueryRequest request = QueryRequestBuilder.New()
+                .SetQuery("query TestQuery($where: FooDateTimeFilter){ foo(where: $where) {foo}}")
+                .SetVariableValue("where", new { foo_gte = "2019-06-01" })
+                .Create();
+
+            // act
+            IExecutionResult result = await executor.ExecuteAsync(request);
+
+            // assert
+            result.MatchSnapshot();
+        }
+
+        public class FooDateTime
+        {
+            public DateTime Foo { get; set; }
+        }
+
+        public class QueryFooDateTime
+        {
+            public IEnumerable<FooDateTime> Foo { get; set; }
         }
 
         public class QueryType
