+++ src/Core/Types/Utilities/TypeConversion.Setup.cs
@@ -55,9 +55,6 @@ namespace HotChocolate.Utilities
             registry.Register<long, DateTime>(
                 from => DateTimeOffset.FromUnixTimeSeconds(from).UtcDateTime);
 
-            registry.Register<string, DateTimeOffset>(
-                from => DateTimeOffset.Parse(from));
-
             registry.Register<DateTimeOffset, string>(
                 from =>
                 {
